+++ FacebookAudienceNetwork/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+ * 5.5.0.0
+    * This version of the adapters has been certified with Facebook Audience Network 5.5.0.
+    * Bidder token generation doesn't depend on Facebook SDK initialization.
+
  * 5.4.1.1
     * Fix banner size passing as part of 5.8.0+ MoPub SDK release changes.